services/geocoding-service/requirements.txt
fastapi[standard]>=0.115.4
duckdb>=1.3.0
pandas>=2.2.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.32.0
pydantic>=2